Add workflow for trusted publishing (#1)
This adds a workflow that will build the bindings for the package along
with the package itself, then publish to PyPi. I've confirmed that the
build works, but publishing isn't working yet due to a problem with the
versioning. It's possible that the problem is that I'm creating a tag
for something that isn't the default branch and that's causing the
version to be a "development" version. I will merge this and tag a new
release on `main` and go from there.